requirements.txt
accounts==0.2.4
alembic==1.13.2
amqp==2.6.1
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
billiard==3.6.4.0
cachetools==5.4.0
celery==4.4.6
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
ecdsa==0.19.0
email-validator==1.3.1
emails==0.5.15
exceptiongroup==1.2.2
fastapi==0.70.1
future==1.0.0
git-filter-repo==2.45.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.7
inflect==5.6.2
inflection==0.5.1
Jinja2==2.11.3
kombu==4.6.11
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
more-itertools==10.4.0
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pluggy==0.13.1
premailer==3.10.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.17
PyJWT==2.9.0
pytest==5.4.3
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2024.1
PyYAML==6.0.2
raven==6.10.0
redis==5.0.8
django-redis==5.0.0
requests==2.32.3
rsa==4.9
six==1.16.0
slack_sdk==3.31.0
sniffio==1.3.1
SQLAlchemy==1.4.53
sqlparse==0.5.1
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.11.8
uvloop==0.19.0
vine==1.3.0
wcwidth==0.2.13
websockets==8.1
whitenoise==6.7.0
